update PPS and PACF so that they also calculate cross correlation terms. Refactored code. Updated tests results files.
