+++ MDANSE/Tests/UnitTests/Analysis/test_grouping.py
@@ -106,8 +106,8 @@ def test_trajectory_state():
         ("VelocityCorrelationFunction", ["vcf"], "equal", 1e-10, 1e-7),
         ("VanHoveFunctionDistinct", ["vh"], "equal", 1e-10, 1e-7),
         ("VanHoveFunctionSelf", ["vh"], "equal", 1e-10, 1e-7),
-        ("PositionAutoCorrelationFunction", ["pacf"], "equal", 1e-10, 1e-7),
-        ("PositionPowerSpectrum", ["pacf", "pps"], "equal", 1e-10, 1e-7),
+        ("PositionCorrelationFunction", ["pcf"], "equal", 1e-10, 1e-7),
+        ("PositionPowerSpectrum", ["pcf", "pps"], "equal", 1e-10, 1e-7),
         ("RootMeanSquareDeviation", ["rmsd"], "equal", 1e-10, 1e-7),
         ("CoordinationNumber", ["cn"], "equal", 1e-10, 1e-7),
         ("PairDistributionFunction", ["pdf", "rdf", "tcf"], "equal", 1e-10, 1e-7),
@@ -258,8 +258,8 @@ def test_ssfsf(generate_benchmarks, tmp_path, dcsf):
 @pytest.mark.parametrize(
     "job_info",
     [
-        ("DensityOfStates", ["dos", "vcf"], "equal", 1e-10, 1e-7),
-        ("PositionPowerSpectrum", ["pacf", "pps"], "equal", 1e-10, 1e-7),
+        ("DensityOfStates", ["dos/isotropic", "vcf/isotropic"], "equal", 1e-10, 1e-7),
+        ("PositionPowerSpectrum", ["pcf/isotropic", "pps/isotropic"], "equal", 1e-10, 1e-7),
         ("DynamicCoherentStructureFactor", ["dcsf"], "b_coherent", 1e-6, 1e-6),
     ],
     ids=lambda x: x[0],
