+++ MDANSE/Tests/UnitTests/Analysis/test_one_atom.py
@@ -25,7 +25,7 @@ ALL_ANALYSES = [
     "GaussianDynamicIncoherentStructureFactor",
     "MeanSquareDisplacement",
     "MolecularTrace",
-    "PositionAutoCorrelationFunction",
+    "PositionCorrelationFunction",
     "RadiusOfGyration",
     "RootMeanSquareDeviation",
     "RootMeanSquareFluctuation",
